Brandon G. Elzy - renamed component file names, installed and set up Redux, and organized file structure
