+++ client/src/components/presentation/presentation.js
@@ -1,0 +1,3 @@
+export { default as CreateAlert     } from './CreateAlert/CreateAlert.js';
+export { default as Topbar          } from './Topbar/Topbar.js';
+export { default as VerticalSidebar } from './VerticalSidebar/VerticalSidebar.js';
