+++ client/src/reducers/reducer.js
@@ -1,0 +1,8 @@
+
+const defaultState = {
+
+};
+
+export const reducer = (state = defaultState, action) => {
+
+}
